requirements.txt
Django==2.2.4
django-debug-toolbar
django-extensions
django-widget-tweaks
djangorestframework
PyJWT
